feat:fixed add to history
