+++ src/App.css
@@ -57,3 +57,9 @@
 		margin: 0;
 	}
 }
+
+@media (max-width: 600px) {
+	.main-container {
+		padding: 8rem 2rem;
+	}
+}
