+++ src/backend/controllers/VideoController.js
@@ -29,10 +29,14 @@ export const updateVideoCountHandler = function (schema, request) {
 	try {
 		const allVideos = this.db.videos;
 		const updatedVideos = allVideos.map((video) => {
-			if (video._id === videoId) video.views += 1;
+			if (video._id === videoId) {
+				console.log(video.views);
+				video.views = video.views + 1;
+				console.log(video.views);
+			}
 			return video;
 		});
-		this.db.videos.update({ videos: { updatedVideos } });
+		this.db.videos.update({ videos: updatedVideos });
 		return new Response(200, {}, { videos: updatedVideos });
 	} catch (error) {
 		return new Response(
